+++ internal/routers/handlers.go
@@ -1,15 +1,11 @@
 package routers
 
 import (
-	"fmt"
-
 	"github.com/gofiber/fiber/v2"
+	"github.com/maktoobgar/go_template/internal/middleware"
+	"github.com/maktoobgar/go_template/internal/services"
 )
 
 func AddRoutes(app *fiber.App) {
-	// GET /api/register
-	app.Get("/api/:name?", func(c *fiber.Ctx) error {
-		msg := fmt.Sprintf("✋ %s", c.Params("name"))
-		return c.SendString(msg)
-	}).Name("temp")
+	app.Group("/api", middleware.Useless).Get(":name?", services.Hi).Name("Hi")
 }