+++ internal/routers/ws.go
@@ -1,26 +1,11 @@
 package routers
 
 import (
-	"log"
-
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/websocket/v2"
+	"github.com/maktoobgar/go_template/internal/services"
 )
 
 func WS(app *fiber.App) {
-	app.Get("/ws", websocket.New(func(c *websocket.Conn) {
-		for {
-			mt, msg, err := c.ReadMessage()
-			if err != nil {
-				log.Println("read:", err)
-				break
-			}
-			log.Printf("recv: %s", msg)
-			err = c.WriteMessage(mt, msg)
-			if err != nil {
-				log.Println("write:", err)
-				break
-			}
-		}
-	}))
+	app.Get("/ws", websocket.New(services.WS))
 }